+++ common/src/main/java/dev/terminalmc/autoreconnectrf/AutoReconnect.java
@@ -13,28 +13,24 @@
 
 package dev.terminalmc.autoreconnectrf;
 
-import com.mojang.logging.LogUtils;
 import com.mojang.realmsclient.RealmsMainScreen;
 import dev.terminalmc.autoreconnectrf.config.Config;
 import dev.terminalmc.autoreconnectrf.reconnect.ReconnectStrategy;
 import dev.terminalmc.autoreconnectrf.reconnect.SingleplayerReconnectStrategy;
 import dev.terminalmc.autoreconnectrf.util.ModLogger;
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.gui.components.Button;
-import net.minecraft.client.gui.components.events.GuiEventListener;
 import net.minecraft.client.gui.screens.DisconnectedScreen;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.client.gui.screens.TitleScreen;
 import net.minecraft.client.gui.screens.multiplayer.JoinMultiplayerScreen;
 import net.minecraft.client.gui.screens.worldselection.SelectWorldScreen;
 import net.minecraft.client.player.LocalPlayer;
-import net.minecraft.network.chat.contents.TranslatableContents;
+import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
-import java.util.Optional;
 import java.util.concurrent.*;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.IntConsumer;
@@ -45,70 +41,19 @@ public class AutoReconnect {
     public static final String MOD_ID = "autoreconnectrf";
     public static final String MOD_NAME = "AutoReconnect-Reforged";
     public static final ModLogger LOG = new ModLogger(MOD_NAME);
-    public static final List<String> DISCONNECT_KEYS = List.of(
-            "disconnect.closed",
-            "disconnect.disconnected",
-            "disconnect.endOfStream",
-            "disconnect.exceeded_packet_rate",
-            "disconnect.genericReason", // arg
-            "disconnect.ignoring_status_request",
-            "disconnect.loginFailed",
-            "disconnect.loginFailedInfo", // arg
-            "disconnect.loginFailedInfo.insufficientPrivileges",
-            "disconnect.loginFailedInfo.invalidSession",
-            "disconnect.loginFailedInfo.serversUnavailable",
-            "disconnect.loginFailedInfo.userBanned",
-            "disconnect.lost",
-            "disconnect.overflow",
-            "disconnect.packetError",
-            "disconnect.spam",
-            "disconnect.timeout",
-            "disconnect.transfer",
-            "disconnect.unknownHost",
 
-            "multiplayer.disconnect.authservers_down",
-            "multiplayer.disconnect.banned",
-            "multiplayer.disconnect.banned_ip.reason", // arg
-            "multiplayer.disconnect.banned.reason", // arg
-            "multiplayer.disconnect.chat_validation_failed",
-            "multiplayer.disconnect.duplicate_login",
-            "multiplayer.disconnect.expired_public_key",
-            "multiplayer.disconnect.flying",
-            "multiplayer.disconnect.generic",
-            "multiplayer.disconnect.idling",
-            "multiplayer.disconnect.illegal_characters",
-            "multiplayer.disconnect.incompatible", // arg
-            "multiplayer.disconnect.invalid_entity_attacked",
-            "multiplayer.disconnect.invalid_packet",
-            "multiplayer.disconnect.invalid_player_data",
-            "multiplayer.disconnect.invalid_player_movement",
-            "multiplayer.disconnect.invalid_public_key_signature",
-            "multiplayer.disconnect.invalid_public_key_signature",
-            "multiplayer.disconnect.invalid_vehicle_movement",
-            "multiplayer.disconnect.ip_banned",
-            "multiplayer.disconnect.kicked",
-            "multiplayer.disconnect.missing_tags",
-            "multiplayer.disconnect.name_taken",
-            "multiplayer.disconnect.not_whitelisted",
-            "multiplayer.disconnect.out_of_order_chat",
-            "multiplayer.disconnect.outdated_client", // arg
-            "multiplayer.disconnect.outdated_server", // arg
-            "multiplayer.disconnect.server_full",
-            "multiplayer.disconnect.server_shutdown",
-            "multiplayer.disconnect.slow_login",
-            "multiplayer.disconnect.too_many_pending_chats",
-            "multiplayer.disconnect.transfers_disabled",
-            "multiplayer.disconnect.unexpected_query_response",
-            "multiplayer.disconnect.unsigned_chat",
-            "multiplayer.disconnect.unverified_username",
-
-            "multiplayer.requiredTexturePrompt.disconnect"
-    );
-
+    // Condition vars
     public static final List<Pattern> conditionPatterns = new ArrayList<>();
-
     public static @Nullable String lastDcReasonStr = null;
     public static @Nullable String lastDcReasonKey = null;
+
+    // Reconnect vars
+    private static final ScheduledThreadPoolExecutor EXECUTOR_SERVICE = new ScheduledThreadPoolExecutor(1);
+    static { EXECUTOR_SERVICE.setRemoveOnCancelPolicy(true); }
+    private static final AtomicReference<ScheduledFuture<?>> countdown = new AtomicReference<>(null);
+    private static @Nullable ReconnectStrategy reconnectStrategy = null;
+
+    // Mod lifecycle methods
 
     public static void init() {
         Config.getAndSave();
@@ -126,67 +71,85 @@ public class AutoReconnect {
         }
     }
 
-    // Legacy
+    // Reconnect methods
 
-    private static final ScheduledThreadPoolExecutor EXECUTOR_SERVICE = new ScheduledThreadPoolExecutor(1);
-    private static final AtomicReference<ScheduledFuture<?>> countdown = new AtomicReference<>(null);
-    private static ReconnectStrategy reconnectStrategy = null;
-
-    static {
-        EXECUTOR_SERVICE.setRemoveOnCancelPolicy(true);
+    /**
+     * Stops any active reconnection, and removes the saved strategy to prevent
+     * future reconnection.
+     *
+     * <p>Any mods wanting to prevent automatic reconnection should invoke this
+     * method at any time after the player has joined a world/server/realm.</p>
+     */
+    public static void cancelAutoReconnect() {
+        cancelActiveReconnect();
+        reconnectStrategy = null;
     }
 
     public static ScheduledFuture<?> schedule(Runnable command, long delay, TimeUnit timeUnit) {
         return EXECUTOR_SERVICE.schedule(command, delay, timeUnit);
     }
 
-    public static void setReconnectHandler(ReconnectStrategy pReconnectStrategy) {
-        if (reconnectStrategy != null) {
-            // should imply that both handlers target the same world/server
-            // we return to preserve the attempts counter
-            assert reconnectStrategy.getClass().equals(pReconnectStrategy.getClass()) &&
-                    reconnectStrategy.getName().equals(pReconnectStrategy.getName());
-            return;
-        }
-        reconnectStrategy = pReconnectStrategy;
-    }
-
-    public static void reconnect() {
-        if (reconnectStrategy == null) return; // shouldn't happen normally, but can be forced
-        cancelCountdown();
-        reconnectStrategy.reconnect();
-    }
-
-    public static void startCountdown(final IntConsumer callback) {
-        // if (countdown.get() != null) return; // should not happen
+    /**
+     * Sets the strategy to be used for the next reconnection attempt.
+     */
+    public static void setReconnectStrategy(@NotNull ReconnectStrategy pReconnectStrategy) {
+        // Avoid overwriting strategy on reconnect failure
         if (reconnectStrategy == null) {
-            // TODO fix issue appropriately, logging error for now
-            LogUtils.getLogger().error("Cannot reconnect because reconnectStrategy is null");
-            callback.accept(-1); // signal reconnecting is not possible
-            return;
+            reconnectStrategy = pReconnectStrategy;
         }
+    }
 
-        int delay = Config.get().getDelayForAttempt(reconnectStrategy.nextAttempt());
+    /**
+     * @return {@code true} if the mod has a reconnection strategy,
+     * {@code false} otherwise.
+     */
+    public static boolean canReconnect() {
+        return reconnectStrategy != null;
+    }
+
+    /**
+     * Attempts to reconnect using the stored strategy.
+     */
+    public static void reconnect() {
+        cancelCountdown();
+        checkStrategy().reconnect();
+    }
+
+    /**
+     * Initiates the countdown for the next reconnect attempt, if any.
+     */
+    public static void startCountdown(final IntConsumer callback) {
+        int delay = Config.get().getDelayForAttempt(checkStrategy().nextAttempt());
         if (delay >= 0) {
             countdown(delay, callback);
         } else {
-            // no more attempts configured
+            // No more attempts configured
             callback.accept(-1);
         }
     }
 
-    public static void cancelAutoReconnect() {
-        if (reconnectStrategy == null) return; // should not happen
-        reconnectStrategy.resetAttempts();
+    /**
+     * Stops attempting reconnection but retains strategy for manual reconnect.
+     */
+    public static void cancelActiveReconnect() {
+        if (reconnectStrategy != null) reconnectStrategy.resetAttempts();
         cancelCountdown();
+    }
+
+    /**
+     * Resets the reconnect countdown and attempts to reconnect using the saved
+     * strategy.
+     */
+    public static void manualReconnect() {
+        AutoReconnect.cancelActiveReconnect();
+        AutoReconnect.reconnect();
     }
 
     public static void onScreenChanged(Screen current, Screen next) {
         if (sameType(current, next)) return;
-        // TODO condition could use some improvement, shouldn't cause any issues tho
+        // TODO condition could use some improvement, shouldn't cause any issues
         if (!isMainScreen(current) && isMainScreen(next) || isReAuthenticating(current, next)) {
             cancelAutoReconnect();
-            reconnectStrategy = null;
         }
     }
 
@@ -258,10 +221,7 @@ public class AutoReconnect {
      */
     private static void sendMessage(LocalPlayer player, String message) {
         if (message.startsWith("/")) {
-            // The first starting slash has to be removed,
-            // otherwise it will be interpreted as a double slash.
-            String command = message.substring(1);
-            player.connection.sendUnsignedCommand(command);
+            player.connection.sendUnsignedCommand(message.substring(1));
         } else {
             player.connection.sendChat(message);
         }
@@ -278,27 +238,17 @@ public class AutoReconnect {
                 screen instanceof JoinMultiplayerScreen || screen instanceof RealmsMainScreen;
     }
 
-    private static boolean isReAuthenticating(Screen from, Screen to) {
-        return from instanceof DisconnectedScreen && to != null &&
-                to.getClass().getName().startsWith("me.axieum.mcmod.authme");
+    private static boolean isReAuthenticating(Screen current, Screen next) {
+        return current instanceof DisconnectedScreen
+                && next != null
+                && next.getClass().getName().startsWith("me.axieum.mcmod.authme");
     }
 
-    public static Optional<Button> findBackButton(Screen screen) {
-        for (GuiEventListener element : screen.children()) {
-            if (!(element instanceof Button button)) continue;
-
-            String translatableKey;
-            if (button.getMessage() instanceof TranslatableContents translatable) {
-                translatableKey = translatable.getKey();
-            } else if (button.getMessage().getContents() instanceof TranslatableContents translatable) {
-                translatableKey = translatable.getKey();
-            } else continue;
-
-            // check for gui.back, gui.toMenu, gui.toRealms, gui.toTitle, gui.toWorld (only ones starting with "gui.to")
-            if (translatableKey.equals("gui.back") || translatableKey.startsWith("gui.to")) {
-                return Optional.of(button);
-            }
+    private static @NotNull ReconnectStrategy checkStrategy() {
+        if (reconnectStrategy == null) {
+            throw new IllegalStateException("Reconnect strategy failed null check");
+        } else {
+            return reconnectStrategy;
         }
-        return Optional.empty();
     }
 }