+++ common/src/main/java/dev/terminalmc/autoreconnectrf/mixin/MixinClientPacketListener.java
@@ -23,7 +23,10 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(ClientPacketListener.class)
 public class MixinClientPacketListener {
-    @Inject(at = @At("TAIL"), method = "handleLogin")
+    @Inject(
+            at = @At("TAIL"),
+            method = "handleLogin"
+    )
     private void onGameJoin(ClientboundLoginPacket packet, CallbackInfo info) {
         AutoReconnect.onGameJoined();
     }