+++ common/src/main/java/dev/terminalmc/autoreconnectrf/mixin/accessor/DisconnectedScreenAccessor.java
@@ -1,5 +1,6 @@
 package dev.terminalmc.autoreconnectrf.mixin.accessor;
 
+import net.minecraft.client.gui.layouts.LinearLayout;
 import net.minecraft.client.gui.screens.DisconnectedScreen;
 import net.minecraft.network.DisconnectionDetails;
 import org.spongepowered.asm.mixin.Mixin;
@@ -9,4 +10,7 @@ import org.spongepowered.asm.mixin.gen.Accessor;
 public interface DisconnectedScreenAccessor {
     @Accessor
     DisconnectionDetails getDetails();
+
+    @Accessor
+    LinearLayout getLayout();
 }