+++ common/src/main/java/dev/terminalmc/autoreconnectrf/util/ScreenMixinUtil.java
@@ -1,0 +1,95 @@
+package dev.terminalmc.autoreconnectrf.util;
+
+import dev.terminalmc.autoreconnectrf.AutoReconnect;
+import dev.terminalmc.autoreconnectrf.config.Config;
+import dev.terminalmc.autoreconnectrf.mixin.accessor.DisconnectedScreenAccessor;
+import net.minecraft.ChatFormatting;
+import net.minecraft.client.gui.components.Button;
+import net.minecraft.client.gui.components.events.GuiEventListener;
+import net.minecraft.client.gui.screens.DisconnectedScreen;
+import net.minecraft.client.gui.screens.Screen;
+import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.contents.TranslatableContents;
+import org.spongepowered.asm.mixin.Unique;
+
+import java.util.Optional;
+import java.util.regex.Pattern;
+
+import static dev.terminalmc.autoreconnectrf.util.Localization.localized;
+
+/**
+ * Yeah, I know it's bad practice.
+ */
+public class ScreenMixinUtil {
+    public static boolean checkConditions(Screen screen) {
+        if (screen instanceof DisconnectedScreen ds) {
+            // Doesn't work with realms disconnect screen
+            Component reason = ((DisconnectedScreenAccessor)ds).getDetails().reason();
+            String reasonStr = reason.getString();
+            AutoReconnect.lastDcReasonStr = reasonStr;
+            AutoReconnect.lastDcReasonKey = null;
+            boolean match = false;
+
+            // Check regex conditions
+            for (Pattern condition : AutoReconnect.conditionPatterns) {
+                if (condition.matcher(reasonStr).find()) {
+                    AutoReconnect.LOG.info("Matched pattern '{}' against reason '{}'",
+                            condition, reasonStr);
+                    match = true;
+                    break;
+                }
+            }
+            // Check key conditions
+            if (!match && reason.getContents() instanceof TranslatableContents tc) {
+                String key = tc.getKey();
+                AutoReconnect.lastDcReasonKey = key;
+                for (String condition : Config.get().options.conditionKeys) {
+                    if (key.contains(condition)) {
+                        AutoReconnect.LOG.info("Matched key '{}' against reason key '{}'",
+                                condition, key);
+                        match = true;
+                        break;
+                    }
+                }
+            }
+            if (Config.get().options.conditionType) {
+                return match;
+            } else {
+                return !match;
+            }
+        }
+        return Config.get().hasAttempts();
+    }
+
+    public static Optional<Button> findBackButton(Screen screen) {
+        for (GuiEventListener element : screen.children()) {
+            if (!(element instanceof Button button)) continue;
+
+            String translatableKey;
+            if (button.getMessage() instanceof TranslatableContents translatable) {
+                translatableKey = translatable.getKey();
+            } else if (button.getMessage().getContents() instanceof TranslatableContents translatable) {
+                translatableKey = translatable.getKey();
+            } else continue;
+
+            // check for gui.back, gui.toMenu, gui.toRealms, gui.toTitle, gui.toWorld (only ones starting with "gui.to")
+            if (translatableKey.equals("gui.back") || translatableKey.startsWith("gui.to")) {
+                return Optional.of(button);
+            }
+        }
+        return Optional.empty();
+    }
+
+    @Unique
+    public static void countdownCallback(Button reconnectButton, int seconds) {
+        if (seconds < 0) {
+            // indicates that we're out of attempts
+            reconnectButton.setMessage(localized("message", "reconnect_failed")
+                    .withStyle(s -> s.withColor(ChatFormatting.RED)));
+            reconnectButton.active = false;
+        } else {
+            reconnectButton.setMessage(localized("message", "reconnect_in", seconds)
+                    .withStyle(s -> s.withColor(ChatFormatting.GREEN)));
+        }
+    }
+}